+++ CHANGELOG.md
@@ -56,6 +56,7 @@ If upgrading from v2.x, see the [v3.0.0 release notes](https://github.com/flixOp
 If upgrading from v2.x, see the [v3.0.0 release notes](https://github.com/flixOpt/flixOpt/releases/tag/v3.0.0) and [Migration Guide](https://flixopt.github.io/flixopt/latest/user-guide/migration-guide-v3/).
 
 ### ✨ Added
+- `overwrite` parameter when saving results to file. If True, overwrite existing files.
 
 ### 💥 Breaking Changes